requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
alembic==1.15.1
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
Flask-APScheduler==1.13.1
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
multidict==6.4.3
mysqlclient==2.2.7
packaging==24.2
paypalrestsdk==1.13.3
propcache==0.3.1
pycparser==2.22
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
stripe==11.6.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.0
